pyproject.toml
[tool.poetry]
name = "kafka-train"
version = "0.1.0"
description = ""
authors = ["Shoma FUKUDA <fkshom@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]
assertpy = "^1.1"
pytest = "^6.2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
